refactor: DI 위치 변경 및 재사용성 구조 변경
