+++ backend/app/api/v1/seller/order.py
@@ -6,7 +6,8 @@ from datetime import datetime, timezone
 
 from utils.docs_error import create_error_responses
 
-from api.deps import CurrentSellerDep, AsyncSessionDep
+from api.deps.auth import CurrentSellerDep
+from api.deps.database import AsyncSessionDep
 from repositories.store import StoreRepository
 from repositories.order_current_item import OrderCurrentItemRepository
 from database.models.order_current_item import OrderCurrentItem
