+++ backend/app/api/v1/seller/product.py
@@ -3,7 +3,8 @@ from fastapi import APIRouter, HTTPException, Depends, status
 
 from utils.docs_error import create_error_responses
 
-from api.deps import CurrentSellerDep, AsyncSessionDep
+from api.deps.auth import CurrentSellerDep
+from api.deps.database import AsyncSessionDep
 from repositories.store import StoreRepository
 from repositories.store_product_info import StoreProductInfoRepository, StockUpdateResult
 from repositories.product_nutrition import ProductNutritionRepository
