+++ backend/app/api/v1/seller/store_images.py
@@ -1,7 +1,8 @@
 from typing import Annotated, List
 from fastapi import APIRouter, HTTPException, Depends, status, UploadFile, File, Path
 from utils.docs_error import create_error_responses
-from api.deps import CurrentSellerDep, AsyncSessionDep
+from api.deps.auth import CurrentSellerDep
+from api.deps.database import AsyncSessionDep
 from repositories.store import StoreRepository
 from repositories.store_image import StoreImageRepository
 from schemas.image import StoreImagesUploadResponse, StoreImagesResponse, ImageUploadResponse
