+++ backend/app/api/v1/seller/store_settings.py
@@ -2,7 +2,8 @@ from typing import Annotated, List
 from fastapi import APIRouter, HTTPException, Depends, status
 
 from utils.docs_error import create_error_responses
-from api.deps import CurrentSellerDep, AsyncSessionDep
+from api.deps.auth import CurrentSellerDep
+from api.deps.database import AsyncSessionDep
 from repositories.store import StoreRepository
 from repositories.address import AddressRepository
 from repositories.store_payment_info import StorePaymentInfoRepository
