+++ backend/app/api/v1/seller/store_sns.py
@@ -3,7 +3,8 @@ from fastapi import APIRouter, HTTPException, Depends, status
 from pydantic import HttpUrl
 
 from utils.docs_error import create_error_responses
-from api.deps import CurrentSellerDep, AsyncSessionDep
+from api.deps.auth import CurrentSellerDep
+from api.deps.database import AsyncSessionDep
 from repositories.store import StoreRepository
 from repositories.store_sns import StoreSNSRepository
 from schemas.store_sns import (
